% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Confidence Interval
return the 95% CI of odds ratio in a string}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficient from logistic regression}

\item{se}{Standard error from logistic regression}

\item{siglevel}{Designated significance level}

\item{roundto}{Decimal places to round to}
}
\value{
ORresult
}
\description{
Confidence Interval
return the 95% CI of odds ratio in a string
}
\examples{
OR_95CI(0,1,0.95,2)
0 (-1.96,1.96)
}
