% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Return the CI of odds ratio}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficient from logistic regression}

\item{se}{Standard error from logistic regression}

\item{siglevel}{Designated significance level}

\item{roundto}{Decimal places to round to}
}
\value{
ORresult
}
\description{
Return the CI of odds ratio
}
\examples{
OR_95CI(0,1,0.95,2)
1.00 (0.94, 1.06)
}
